Refactor : 패키지 이름 변경
- ladder -> domain
- User 객체를 만들기 위해 패키지 이름 변경
