+++ src/main/java/kr/codesquad/domain/Ladder.java
@@ -1,0 +1,20 @@
+package kr.codesquad.domain;
+
+import java.util.List;
+
+public class Ladder {
+    private LadderMaker ladderMaker;
+    private List<String> users;
+
+    public Ladder() {
+
+    }
+
+    public void run() {
+
+    }
+
+
+
+
+}
